LPS-72096 Move tests that can be moved, to asset-tests module, no logical changes
