+++ modules/apps/web-experience/asset/asset-test/src/testIntegration/java/com/liferay/asset/assetrendererfactoryregistryutil/test/AssetRendererFactoryRegistryUtilTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.asset;
+package com.liferay.asset.assetrendererfactoryregistryutil.test;
 
 import com.liferay.asset.kernel.AssetRendererFactoryRegistryUtil;
 import com.liferay.asset.kernel.model.AssetRendererFactory;
@@ -20,7 +20,6 @@ import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.SyntheticBundleRule;
 import com.liferay.portal.util.PortalImpl;
-import com.liferay.portlet.asset.bundle.assetrendererfactoryregistryutil.TestAssetRendererFactory;
 
 import java.util.List;
