+++ modules/apps/web-experience/asset/asset-test/src/testIntegration/java/com/liferay/asset/assetrendererfactoryregistryutil/test/bnd.bnd
