+++ modules/apps/web-experience/asset/asset-test/src/testIntegration/java/com/liferay/asset/search/test/AssetVocabularySearchTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.asset.search;
+package com.liferay.asset.search.test;
 
 import com.liferay.asset.kernel.model.AssetVocabulary;
 import com.liferay.asset.kernel.service.AssetVocabularyLocalServiceUtil;
