+++ modules/apps/web-experience/asset/asset-test/src/testIntegration/java/com/liferay/asset/service/test/AssetTagLocalServiceTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.asset.service;
+package com.liferay.asset.service.test;
 
 import com.liferay.asset.kernel.model.AssetTag;
 import com.liferay.asset.kernel.model.AssetTagStats;
