+++ modules/apps/web-experience/asset/asset-test/src/testIntegration/java/com/liferay/asset/service/test/TestAssetIndexer.java
@@ -12,10 +12,29 @@
  * details.
  */
 
-package com.liferay.portlet.asset.bundle.assetrendererfactoryregistryutil;
+package com.liferay.asset.service.test;
+
+import com.liferay.portlet.usersadmin.util.OrganizationIndexer;
+
+import org.junit.Assert;
 
 /**
- * @author Mate Thurzo
+ * @author Michael C. Han
  */
-public class TestAssetRendererModel {
+public class TestAssetIndexer extends OrganizationIndexer {
+
+	@Override
+	public void reindex(String className, long classPK) {
+		Assert.assertEquals(_className, className);
+		Assert.assertEquals(_classPK, classPK);
+	}
+
+	public void setExpectedValues(String className, long classPK) {
+		_className = className;
+		_classPK = classPK;
+	}
+
+	private String _className;
+	private long _classPK;
+
 }
